add jtoday.sh
jtoday is a journaling tool that is basically a shortcut to open a file of
the format YYYY-mm-dd-journal.txt with vim

I'll be using this as an easy way to keep a plain text journal in
Dropbox. I tried evernote per Azeem's rec, but had syncing issues. I
like the idea of maintianing a journal, however, so rather than using
evernote, I'll use tools I trust: Dropbox, Unix, vim.